+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Added sort options to GSEA bubble.
 - GSEA bubble now has own app.
 - Moved gsea bubble into own app.
 - GSEA dot now has refined tooltips.